requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
config==0.5.1
Django==4.1.2
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
google-auth==2.17.1
idna==3.4
pg==0.1
Pillow==9.2.0
psycopg-binary==3.1.4
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.6.0
PyOpenGL==3.1.6
pytz==2022.5
requests==2.28.2
rsa==4.9
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
